fix: split into components
