+++ packages/shared/src/features/opportunity/components/analyze/RelevantBlock.tsx
@@ -1,0 +1,165 @@
+import React, { useEffect, useState } from 'react';
+import {
+  Typography,
+  TypographyColor,
+  TypographyType,
+} from '../../../../components/typography/Typography';
+import { Divider } from '../../../../components/utilities';
+import { Chip } from '../../../../components/cards/common/PostTags';
+import { Loader } from '../../../../components/Loader';
+import { useOpportunityPreviewContext } from '../../context/OpportunityPreviewContext';
+import { cloudinarySquadsImageFallback } from '../../../../lib/image';
+
+const iconSize = 24;
+
+type CompanyItemProps = {
+  favicon?: string;
+  name: string;
+  apiUrl: string;
+};
+
+const CompanyItem = ({ favicon, name, apiUrl }: CompanyItemProps) => (
+  <div className="flex gap-2">
+    <img
+      src={`${apiUrl}/icon?url=${encodeURIComponent(favicon)}&size=${iconSize}`}
+      className="size-4 rounded-2"
+      alt={name}
+    />
+    <Typography type={TypographyType.Footnote} color={TypographyColor.Tertiary}>
+      {name}
+    </Typography>
+  </div>
+);
+
+type RelevantBlockProps = {
+  loadingStep: number;
+  apiUrl: string;
+};
+
+export const RelevantBlock = ({ loadingStep, apiUrl }: RelevantBlockProps) => {
+  const data = useOpportunityPreviewContext();
+  const totalCount = data?.result?.totalCount ?? 0;
+  const tags = data?.result?.tags ?? [];
+  const companies = data?.result?.companies ?? [];
+  const squads = data?.result?.squads ?? [];
+  const hasData = totalCount > 0 || tags.length > 0;
+  const [animatedCount, setAnimatedCount] = useState(0);
+  const [showSubtext, setShowSubtext] = useState(false);
+
+  // Show after step 2 completes (Mapping skills, requirements, and intent)
+  const isVisible = hasData && loadingStep >= 2;
+
+  // Animate count from 0 to totalCount
+  useEffect(() => {
+    if (!isVisible || totalCount === 0) {
+      return undefined;
+    }
+
+    const duration = 1000; // 1 second animation
+    const steps = 30;
+    const increment = totalCount / steps;
+    const stepDuration = duration / steps;
+    let currentStep = 0;
+
+    const timer = setInterval(() => {
+      currentStep += 1;
+      if (currentStep >= steps) {
+        setAnimatedCount(totalCount);
+        setShowSubtext(true);
+        clearInterval(timer);
+      } else {
+        setAnimatedCount(Math.floor(increment * currentStep));
+      }
+    }, stepDuration);
+
+    return () => clearInterval(timer);
+  }, [isVisible, totalCount]);
+
+  if (!isVisible) {
+    return (
+      <div className="flex items-center justify-center py-8">
+        <Loader />
+      </div>
+    );
+  }
+
+  return (
+    <div className="animate-fade-in flex flex-col gap-4">
+      <div className="flex flex-col gap-1">
+        <Typography type={TypographyType.Body} bold>
+          Relevant active developer
+        </Typography>
+        <Typography
+          type={TypographyType.Title2}
+          bold
+          color={TypographyColor.Brand}
+        >
+          {animatedCount.toLocaleString()}
+        </Typography>
+        {showSubtext && (
+          <Typography
+            type={TypographyType.Callout}
+            color={TypographyColor.Tertiary}
+          >
+            30% searching for job only on daily.dev
+          </Typography>
+        )}
+      </div>
+      <Divider className="bg-border-subtlest-tertiary" />
+      <Typography type={TypographyType.Footnote} bold>
+        Interesting in
+      </Typography>
+      <div className="flex flex-wrap gap-2">
+        {tags.map((tag) => (
+          <Chip key={tag} className="!my-0 !text-text-tertiary">
+            {tag}
+          </Chip>
+        ))}
+      </div>
+      <Divider className="bg-border-subtlest-tertiary" />
+      <Typography type={TypographyType.Footnote} bold>
+        Developers verified work at
+      </Typography>
+      <div className="grid grid-cols-2 gap-2">
+        {companies.map((company) => (
+          <CompanyItem
+            key={company.name}
+            favicon={company.favicon}
+            name={company.name}
+            apiUrl={apiUrl}
+          />
+        ))}
+      </div>
+      <Divider className="bg-border-subtlest-tertiary" />
+      <div>
+        <Typography type={TypographyType.Footnote} bold>
+          Most active squads
+        </Typography>
+        <Typography
+          type={TypographyType.Footnote}
+          color={TypographyColor.Tertiary}
+        >
+          Where your matches are most active inside daily.dev.
+        </Typography>
+      </div>
+      <div className="grid grid-cols-2 gap-2">
+        {squads.map((squad) => (
+          <div className="flex gap-2" key={squad.handle}>
+            <img
+              src={cloudinarySquadsImageFallback}
+              className="size-4 rounded-full"
+              alt={squad.image}
+            />
+            <Typography
+              type={TypographyType.Footnote}
+              color={TypographyColor.Tertiary}
+            >
+              {squad.handle}
+            </Typography>
+          </div>
+        ))}
+      </div>
+      <Divider className="bg-border-subtlest-tertiary" />
+    </div>
+  );
+};
